+++ src/utils/createSvgHtml.ts
@@ -1,52 +1,42 @@
 import type { IconType } from './icons'
 
-export const createPaths = (paths: string[], jsx: boolean, type: IconType) => (type === 'outline'
-  ? paths
-    .reduce((content, path) => `
-      ${content}
-      <path fill${jsx ? 'R' : '-r'}ule="evenodd" d="${path}" fill="currentColor" />
-    `, '')
-  : paths
-    .reduce((content, path) => `
-      ${content}
-      <path d="${path}" stroke="currentColor" stroke${jsx ? 'L' : '-l'}inecap="round" stroke${jsx ? 'L' : '-l'}inejoin="round" />
-    `, '')
-)
-
-type HtmlOptionsClasses = {
-  classes: true
-  className: string
-  color?: string
-  size?: string
-}
-
-type HtmlOptionsNoClasses = {
-  classes: false
-  className?: string
-  color: string
+type StyleOptions = {
   size: string
+  color: string
 }
 
-const createModifiableAttrsHTML = ({
-  classes, className, color, size,
-}: HtmlOptionsClasses | HtmlOptionsNoClasses) => (classes ? `class="${className}"` : `width="${size}" height="${size}" style="color: ${color}"`)
-
-const createModifiableAttrsJSX = ({
-  classes, className, color, size,
-}: HtmlOptionsClasses | HtmlOptionsNoClasses) => (
-  classes
-    ? `className="${className}"`
-    : `width="${size}" height="${size}" style="{{ color: ${color} }}"`
+const createStrokePaths = (paths: string[], jsx: boolean) => (
+  paths.reduce((content, path) => `
+    ${content}
+    <path d="${path}" stroke="currentColor" stroke${jsx ? 'L' : '-l'}inecap="round" stroke${jsx ? 'L' : '-l'}inejoin="round" />
+  `, '')
 )
 
-export const createSvgHtml = (svgOptions: HtmlOptionsClasses | HtmlOptionsNoClasses) => (`
-  <svg ${createModifiableAttrsHTML(svgOptions)} viewBox="0 0 24 24" fill="none">
-    {paths}
-  </svg>
-`)
+const createFillPaths = (paths: string[], jsx: boolean) => (
+  paths.reduce((content, path) => `
+    ${content}
+    <path fill${jsx ? 'R' : '-r'}ule="evenodd" d="${path}" fill="currentColor" />
+  `, '')
+)
 
-export const createSvgJsx = (svgOptions: HtmlOptionsClasses | HtmlOptionsNoClasses) => (`
-  <svg ${createModifiableAttrsJSX(svgOptions)} viewBox="0 0 24 24" fill="none">
-    {paths}
-  </svg>
-`)
+export const createPaths = (paths: string[], jsx: boolean, type: IconType) => (
+  type === 'stroke'
+    ? createStrokePaths(paths, jsx)
+    : createFillPaths(paths, jsx)
+)
+
+const setNonClassAttrs = (jsx: boolean, { size, color }: StyleOptions) => (
+  `style=${jsx ? `"{{ color: ${color} }}"` : `"color: ${color}"`} width="${size}" height="${size}"`
+)
+const setClassAttr = (jsx: boolean, className: string) => `class${jsx ? 'Name' : ''}="${className}"`
+
+export const createSvg = <T extends boolean>(
+  usingClasses: T,
+  options: T extends true ? { className: string } : StyleOptions,
+  jsx: boolean,
+) => (
+    // @ts-ignore
+    `<svg ${usingClasses ? setClassAttr(jsx, options.className) : setNonClassAttrs(jsx, options)} viewBox="0 0 24 24" fill="none">
+      {paths}
+    </svg>`
+  )